fixed the reddit web scraper to support the changed reddit HTML for the drawing prompt cog.
reorganized cogs to support a more intelligent file organization. Also updated the docstrings to support reStructuredText format.
